requirements.txt
# Python 3.12.12
# For GPU (CUDA 12.1): pip install torch==2.2.2 torchvision==0.17.2 --index-url https://download.pytorch.org/whl/cu121

# AI & Computer Vision
torch==2.2.2
torchvision==0.17.2
timm==1.0.22
facenet-pytorch==2.5.3

# Transformers
transformers==4.57.3

# Data Science & Visualization
pandas==2.2.2
numpy==1.26.4
matplotlib==3.10.0
seaborn==0.13.2
scikit-learn==1.6.1

# Others
pillow==10.2.0
tqdm
gradio